Update index.rst
The "the" before "MIT License" is required for the *Be Safe* sentence to read correctly. Instead of referring to "our software", I refer to "your project" as in the other points. I also removed the reference to plural projects in the *Promote* point as all other points refer to a project in the singular.